requirements.txt
azure-identity
azure-mgmt-compute
azure-mgmt-resource
boto3
botocore
cryptography>=42.0.5
jinja2>=3.1.3
paramiko>=3.4.0
progressbar2
pyyaml
requests>=2.31.0
requests[security]>=2.31.0
requests-toolbelt
s3transfer
scp
urllib3
